requirements.txt
pandas
numpy
re
json
requests
sklearn.metrics
sklearn.model_selection
torch
transformers
datasets
sentence-transformers
trl
vllm
unsloth